API

ClosedWaveguideDispersion.assemble_A

ClosedWaveguideDispersion.assemble_B

ClosedWaveguideDispersion.calc_diagram

ClosedWaveguideDispersion.plot_diagram

ClosedWaveguideDispersion.setup_bdcs

ClosedWaveguideDispersion.setup_grid

ClosedWaveguideDispersion.assemble_A — Method

assemble_A(cv::CellValues, dh::DofHandler, A::SparseMatrixCSC, α::Float64)

TBW

ClosedWaveguideDispersion.assemble_B — Method

assemble_B(cv::CellValues, dh::DofHandler, B::SparseMatrixCSC, n::Function)

TBW

ClosedWaveguideDispersion.calc_diagram — Method

calc_diagram(cv::CellValues, dh::DofHandler, cst::ConstraintHandler, A::SparseMatrixCSC, B::SparseMatrixCSC, n::Function, bz; nevs::Int=6)

TBW

ClosedWaveguideDispersion.plot_diagram — Method

plot_diagram(bz, μ;period=1.0, title="Dispersion Diagram", xlabel="Brillouin zone")

TBW

ClosedWaveguideDispersion.setup_bdcs — Method

setup_bdcs(dh::DofHandler; period=1.0)

Impose the periodic boundary condition on the two vertical boundaries on the periodic cell.

ClosedWaveguideDispersion.setup_grid — Method

setup_grid(;lc=0.05, period=1.0, height=1.0)

Generate the mesh for the periodic cell by using Gmsh.

Arguments

lc: the mesh size

period: the period of the periodic closed waveguide

height: the height of the periodic closed waveguide